+++ services/agents/activities/index.ts
@@ -1,5 +1,6 @@
 export { calculateAgentTimeUsageMs } from "./agent-time-usage";
 export { AGENT_TIME_CHARGE_LIMIT_MINUTES } from "./constants";
+export { AgentActivity } from "./deprecated-agent-activity";
 export { hasEnoughAgentTimeCharge } from "./has-enough-agent-time-charge";
-export { AgentActivity } from "./types";
+export { saveAgentActivity } from "./save-agent-activity";
 export { getMonthlyBillingCycle } from "./utils";
